removes the meta version of the contact RichData
This block was not going to work, here is the test output of the old
block.
http://www.google.com/webmasters/tools/richsnippets?q=https%3A%2F%2Fraw.github.com%2FjeremyBass%2FWSU-spine%2Fmaster%2Fspine.html
and the current block,
http://www.google.com/webmasters/tools/richsnippets?q=https%3A%2F%2Fraw.github.com%2FjeremyBass%2FWSU-spine%2Fmaster%2Ftest%2Ftest.html
.  Also there are comments in the html above the new corrected example
block.. This block is only an example as the attributes are what we will
be drawing for the look up as we do the html generation in javascript
for the contact tab.