Introduce pkg.build_version, use grunt templating for version output
This removes the declared version var at the top of gruntfile.js and uses package.json to pull build version information when creating a build directory.

By having a build_version of something like 0.5, we're able to maintain smaller versions of 0.5.x to indicate revision while not causing a full URL change.