Refactor sass global locality and introduce font_domain
Locality was hard coded to spine version 1 and not used everywhere.

I updated locality to point to the current develop version. We'll
need to determine the best way to track this for release going
forward. If we can pass in something from grunt, that would be ideal
so that we aren't jumping around looking for versions to bump.

Locality can always be https, because loading assets via https is
not a bad thing.

font_domain requires a relative URL as browsers have issues with
CORs requests, even across scheme