An initial attempt at a CONTRIBUTING document
This CONTRIBUTING document can help describe our development
workflow to anyone submitting changes to the project.

I've gone with a basic description of our new branching system and
some getting started links for creating patches through forking
and branching. It's probably best to lean toward a trim page here,
but it is nice to have as much relevant information as possible.