Consolidate SVG support logic into the `svg_enabled` check
Previously, we would look for all SVG images in an lt-ie9
environment and replace them with their PNG equivalents. The check
attached to this was misspelled, so replacement would never have
happened.

We can consolidate this with the check for SVG support and only
do this check when we know the browser does not support SVGs