Introduce `.spine-mobile`, `.spine-full`, and `.spine-mobile-open`
`.spine-mobile` and `.spine-full` are attached to `html` and
indicate what type of Spine is being implemented on this view. For
the time being, this is a detection for iOS and Android devices.
This can be expanded to use screen size detection once things are
more stable.

`.spine-mobile-open` is attached to `body` when the Spine nav is
shown on a small screen. This helps control the display of the
nav area.

The concepts of "shelved" and "unshelved" are retired as of this
commit. Switching to higher level classes allows us some more
control over the stack.

On a mobile view, `.spine-header` is moved out of `#glue` so that
it can retain its true place as a fixed header without worrying
about the impacts of other transition effects.