Cleanup inline SCSS comments to avoid empty CSS rules
In SASS, comments with `/* */` are kept in the final CSS output
while comments with `// ` are removed from the final output. The
layout of several rules/comments in our current setup results in
empty rules being output in the final stylesheet. We can adjust
these inline comments so that they aren't output in the final file.

We can also now turn on the empty rules config in csslint.