Reduce repeated declarations of WSU Signature marks (stage 1)
Rather than repeating each loop of signatures for lighter, darker,
etc... only to assign the same two marks (white/regular), we can
reduce the logic a bunch to apply the same background to many
styles of the signature.

Really, we should be able to remove the specificity entirely at
some point, but that may come in a later stage.

This already saves a bunch of space on the minified CSS