Use `window.matchMedia` to determing current media query state
Rather than fight with innerWidth, outerWidth, and the various
ways that browsers report things, we can use the `matchMedia` API
to determine what media query is being used. This allows us to
easily match with the media queries in our stylesheet when moving
between mobile and full views of the Spine.

IE 8 and IE 9 do not support `matchMedia`, but do support their
`styleMedia`. We can use this as a fallback for IE 9 and let IE 8
do its thing as it does not support media queries anyway. If you
find yourself worried about the mobile view of IE8, there are
likely other problems.