Remove the special `.fixed` handling between 694px and 791px
There was an attempt here to have the non-mobile navigation appear
next to a smaller `main`, but it was not fully fleshed out and/or
has been broken for a while.

We can let things handle themselves with our other media queries
for max width 693px and max width 791px instead.