Remove extra menu SVG icons
These have never been used. The menu.svg and menu-white.svg are
also not used in the current version of the Spine, and were
removed in #275 from the stylesheet.

There are still a small number of sites running old versions of
the Spine that request these files, so I'm going to leave it
in the repository for the time being until we can get that nubmer
lower.