Remove attempt at markdown template support in built test html
The sidebar.md file that we're attempting to use with nunjucks
markdown is not actually included in any of the built files, so
something is broken in the process.

Rather than fix it, we can remove all of this markdown handling
in favor of the current templating we have going on. The links
in that sidebar.md aren't relevant to the Spine repository.