Remove template logic associated with building spine.*html files
We provide the canonical "spine.html" and "spine.min.html" at the
root of the repository. This build process would create a file
that was then overwritten by another Grunt process. That has been
cleared in an earlier commit and we can now remove the template
logic associated with it.