Simplify and document scripting to add `active`, `dogeared` to nav
`active` and `dogeared` are both important for proper styling in
the navigation. This clarifies some of the logic attached to applying
these styles and drops a `.active:not(:has(.active))` that was super
confusing.

From what I can tell, everything appears to work as is and better
in our test suite, which uses the `active` method and relies on
the Spine to add `dogeared`.