Reset the `expr` option in `jshint` to default
From jshint docs:

"This option suppresses warnings about the use of expressions where normally you would expect to see assignments or function calls. Most of the time, such code is a typo. However, it is not forbidden by the spec and that's why this warning is optional."

This does not require any code changes in the Spine