Calculate Spine action tab heights on window height, not main
On page views where `main` was longer than the navigation, the
min height attached to the Spine action when an action was clicked
would be as high as main and create a confusing experience. This
change locks it to the height of the browser, similar to how the
Spine itself is handled.

Fixes #291