Process couplets into Overview anchors before checking active state
By reversing this logic a bit, we can avoid the cloning of an item
that has already been determined as active. Instead, an Overview
item will be cloned as current/active only if the original copy
had those classes assigned and not as a side effect of another
child item sending the active class up the chain.